[fastmanifest] reduce the warning spew from mercurial build
Summary:
`setup.py` turns on a couple extra warnings, so add them to our list and resolve them.  Most of them are 64->32 bit conversions.  In most cases, we test and abort if we're outside the 32-bit range.

I also removed const from `contains_path` because enforcing it means code duplicating, which is never a great thing.

Test Plan: pass unit tests

Reviewers: #sourcecontrol, rmcelroy

Reviewed By: rmcelroy

Subscribers: mitrandir, mjpieters

Differential Revision: https://phabricator.fb.com/D3194045

Tasks: 10652171

Signature: t1:3194045:1461605953:1a7012fbc75ca8808130c8ad00ed4b2e8fa96f91

+++ fastmanifest/bsearch_test.c
@@ -7,7 +7,7 @@
 #include "bsearch.h"
 #include "tests.h"
 
-#define CMP(left, right) (*((intptr_t*) left) - *((intptr_t*) right))
+#define CMP(left, right) ((int) (*((intptr_t*) left) - *((intptr_t*) right)))
 
 COMPARATOR_BUILDER(intptr_cmp, CMP)
